Fixed product list generation performance issue, moved types into main.d.ts, ProductList implemented
