+++ src/components/ProductCard.tsx
@@ -1,11 +1,15 @@
 import React from 'react';
 
 interface Props {
-
+  cardData: productType
 }
 
 class ProductCard extends React.Component<Props> {
-
+  render() {
+    return (
+      <div>placeholder text!!</div>
+    )
+  }
 }
